libri_speech/requirements.txt
pydub
librosa
numpy
soundfile
scipy
openai-whisper
nltk
scikit-learn
sentence-transformers
torch
